refactor: move usePagination to shared area and switched to compat router
